BUG: Histogramdd breaks on big arrays in Windows (#22561)
* BUG: Histogramdd breaks on big arrays in Windows

Resolved the issue with line change from int to np.intp in numpy/numpy/lib/histograms.py

* BUG: Histogramdd breaks on big arrays in Windows

Resolved the issue with line change from int to np.intp in numpy/numpy/lib/histograms.py

* Removed the binary files

* Update test_histograms.py

* Update test_histograms.py

* Update test_histograms.py

+++ numpy/lib/histograms.py
@@ -1019,7 +1019,7 @@ def histogramdd(sample, bins=10, range=None, normed=None, weights=None,
         sample = np.atleast_2d(sample).T
         N, D = sample.shape
 
-    nbin = np.empty(D, int)
+    nbin = np.empty(D, np.intp)
     edges = D*[None]
     dedges = D*[None]
     if weights is not None:
